requirements.txt
discord.py==1.6.0
python-dotenv==0.15.0
IMDbPY==2020.9.25
gspread==3.7.0
oauth2client==4.1.3